+++ src/ExifTool.spec.ts
@@ -1,10 +1,11 @@
 import { BatchCluster } from "batch-cluster"
 import * as _path from "path"
 
-import { expect } from "./_chai.spec"
+import { expect, testImg } from "./_chai.spec"
 import { times } from "./Array"
 import { DefaultMaxProcs, ExifTool, exiftool } from "./ExifTool"
 import { Tags } from "./Tags"
+import { WriteTask } from "./WriteTask"
 
 describe("ExifTool", () => {
   const et = new ExifTool({ maxProcs: 2 })
@@ -199,9 +200,36 @@ describe("ExifTool", () => {
       secondTasks.forEach(t => expect(t).to.not.be.undefined)
       await et.end()
       expect(await et.pids).to.eql([])
-      return
     } finally {
       await et.end()
     }
+    return
+  })
+
+  // This test failed in versions before
+  it("handles warnings properly", async function() {
+    this.slow(1000)
+    const et = new ExifTool()
+    try {
+      const img = await testImg("bad-exif-ifd.jpg")
+      // Hey, if you're reading this--this isn't how you should normally do
+      // writes. Use ExifTool.write().
+
+      // This test needs access to the errors array, so I need to keep
+      // it in scope (so I can't use exiftool.write()).
+      const tasks: WriteTask[] = []
+      await et.enqueueTask(() => {
+        const t = WriteTask.for(img, { AllDates: new Date().toISOString() })
+        tasks.push(t)
+        return t
+      })
+      const lastTask = tasks[tasks.length - 1]!
+      expect(lastTask.errors[0]).to.match(
+        /Warning: Duplicate MakerNoteUnknown tag in ExifIFD/
+      )
+    } finally {
+      await et.end()
+    }
+    return
   })
 })