+++ src/ExifToolTask.ts
@@ -1,21 +1,19 @@
 import * as bc from "batch-cluster"
 
-export function cmd(args: string[]): string {
-  return [...args, "-ignoreMinorErrors", "-execute", ""].join("\n")
-}
-
 export abstract class ExifToolTask<T> extends bc.Task<T> {
+  static renderCommand(args: string[]): string {
+    return [...args, "-ignoreMinorErrors", "-execute", ""].join("\n")
+  }
+
   readonly errors: string[] = []
 
-  protected constructor(args: string[]) {
-    super(cmd(args), data => this.parse(data))
+  constructor(args: string[]) {
+    super(ExifToolTask.renderCommand(args), data => this.parse(data))
   }
 
   protected abstract parse(input: string): T
 
-  public addError(error: any) {
-    if (error != null) {
-      this.errors.push(String(error))
-    }
+  addError(err: string) {
+    this.errors.push(err)
   }
 }